pyproject.toml
[project]
name = "trd_ticker"
version = "0.1.0"
description = ""
authors = [
    {name = "Aulia Widyaputra"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "google-adk (>=1.21.0,<2.0.0)",
    "aiohttp[speedups] (>=3.13.2,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "agent_ticker"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)"
]
